clean up readme, better log print, show memory usage
