--- .vscode/tasks.json
@@ -1,20 +1,0 @@
-{
-	"version": "2.0.0",
-	"tasks": [
-    {
-      "label": "eslint (node task)",
-      "command": "node",
-      "presentation": { "echo": false, "focus": true, "showReuseMessage": true, "clear": true },
-      // we could've used "$eslint-stylish", but I'm not sure if I like the problems panel that much
-      "problemMatcher": [],
-      "args": ["node_modules/eslint/bin/eslint", "--format=unix", "src"],
-      "group": { "kind": "build", "isDefault": true }
-    },
-    {
-      "label": "eslint fix (npm task)",
-      "type": "npm",
-      "script": "lint:fix",
-      "problemMatcher": []
-    }
-  ]
-}
