+++ package.json
@@ -1,6 +1,6 @@
 {
   "name": "springald",
-  "version": "1.0.1",
+  "version": "1.0.2",
   "description": "Simple application launcher, similar to Launchy.",
   "author": "Szabolcs Kurdi",
   "license": "ISC",
