+++ src/shared/log.ts
@@ -31,6 +31,31 @@ const _log = (action: ILogAction = 'log', ...args: any[]) => {
   _rawLog(action, ...args);
 };
 
+const prettyPrintBuffers = (logs: Record<string, ILogBuffer>): void => {
+  // do NOT use the log wrapper here, we don't want this end up in the buffer
+  for (const section of Object.keys(logs)) {
+    _rawLog('log', '%c' + section.toUpperCase(), 'font-weight:bold;color:purple');
+    for (const group of Object.keys(logs[section])) {
+      const bufferKey = group as ILogAction;
+      const buffers: ILogBuffer = logs[section];
+      const buffer = buffers[bufferKey];
+      if (buffer.length === 0) {
+        continue;
+      }
+      for (const logEvent of buffer) {
+        const restArgs = logEvent.args.slice(1);
+        const at = new Date(logEvent.at).toLocaleString('sv-SE', { hour12: false });
+        const text = bufferKey.padEnd(6, ' ') + '| ' + at + ' | ' + logEvent.args[0];
+        if (restArgs.length > 0) {
+          _rawLog('log', text, restArgs);
+        } else {
+          _rawLog('log', text);
+        }
+      }
+    }
+  }
+};
+
 export const log = {
   debug: (...args: any[]) => _log('debug', ...args),
   info: (...args: any[]) => _log('info', ...args),
@@ -43,6 +68,7 @@ export const log = {
   noBufferLog: (...args: any[]) => _rawLog('log', ...args),
   noBufferWarn: (...args: any[]) => _rawLog('warn', ...args),
   noBufferError: (...args: any[]) => _rawLog('error', ...args),
+  prettyPrint: prettyPrintBuffers,
   // ---
   getBuffer: () => buffer, // be careful, this is NOT a clone
   getErrorCount: (inputBuffer: ILogBuffer) => (inputBuffer || buffer).error.length,
